python_scripts/requirements.txt
amqp==5.2.0
appnope==0.1.3
asttokens==2.4.1
billiard==4.2.1
blinker==1.8.2
camelot-py==0.11.0
celery==5.4.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
configparser==7.1.0
cryptography==43.0.3
decorator==5.1.1
et-xmlfile==1.1.0
excalibur-py==0.4.3
exceptiongroup==1.1.3
executing==2.0.1
Flask==3.0.3
fuzzywuzzy==0.18.0
ghostscript==0.7
greenlet==3.1.1
importlib_metadata==8.5.0
ipdb==0.13.13
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
kombu==5.4.2
Levenshtein==0.26.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
numpy==2.0.0
opencv-python==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
parso==0.8.3
pdfminer.six==20240706
pdftopng==0.2.3
pexpect==4.8.0
pip-autoremove==0.10.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pycrypto==2.6.1
Pygments==2.16.1
pypdf==5.0.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.0
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.0
six==1.16.0
SQLAlchemy==2.0.36
stack-data==0.6.3
tabulate==0.9.0
tomli==2.0.1
traitlets==5.13.0
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.9
Werkzeug==3.0.4
xlrd==2.0.1
zipp==3.20.2
